+++ crates/calib-targets-chessboard/src/boosters.rs
@@ -7,7 +7,7 @@
 //!
 //! The structural skeleton (cell enumeration, KD-tree, per-cell
 //! attachment ladder, fixed-point iteration) lives in
-//! [`projective_grid::seed_and_grow::fill::fill_grid_holes`]; this module
+//! [`projective_grid::shared::fill::fill_grid_holes`]; this module
 //! wraps it with a chessboard-specific [`SquareAttachPolicy`] that adds:
 //!
 //! - **Weak-cluster rescue**: admit `NoCluster` corners whose
@@ -26,10 +26,8 @@ use crate::grow::GrowResult;
 use crate::params::DetectorParams;
 use calib_targets_core::AxisEstimate;
 use nalgebra::Point2;
-use projective_grid::seed_and_grow::fill::{fill_grid_holes, FillParams};
-use projective_grid::seed_and_grow::grow::{
-    Admit, FillEdgeCtx, LabelledNeighbour, SquareAttachPolicy,
-};
+use projective_grid::shared::fill::{fill_grid_holes, FillParams};
+use projective_grid::shared::grow::{Admit, FillEdgeCtx, LabelledNeighbour, SquareAttachPolicy};
 use std::collections::{HashMap, HashSet};
 
 /// Diagnostic returned by the recall-booster stage.
