+++ crates/calib-targets-chessboard/src/grow.rs
@@ -2,10 +2,10 @@
 //!
 //! The pattern-agnostic [`GrowResult`] container — the `(i, j) → index`
 //! labelling plus the cardinal step vectors and rebase parity — lives in
-//! [`projective_grid::seed_and_grow::grow`]. The chessboard crate re-exports
-//! it here so the topological recovery, boosters, geometry check, and output
-//! stages share one labelled-grid type.
+//! [`projective_grid::shared::grow`]. The chessboard crate re-exports it here
+//! so the topological recovery, boosters, geometry check, and output stages
+//! share one labelled-grid type.
 
-use projective_grid::seed_and_grow::grow as pg_grow;
+use projective_grid::shared::grow as pg_grow;
 
 pub use pg_grow::GrowResult;
