+++ crates/calib-targets-chessboard/src/params/advanced.rs
@@ -200,7 +200,7 @@ fn default_enable_post_grow_bfs_regrow() -> bool {
 fn default_enable_post_grow_bfs_extend() -> bool {
     // Default ON. After refit produces refined centres, walk the
     // existing labelled set's boundary with a non-destructive BFS
-    // (`projective_grid::seed_and_grow::grow_extend::extend_from_labelled`),
+    // (`projective_grid::shared::grow_extend::extend_from_labelled`),
     // attaching newly-Clustered corners via cardinal-neighbour
     // propagation. Reaches interior-hole / left-strip corners that
     // local-H extrapolation (`extend_boundary` /
@@ -536,7 +536,7 @@ pub struct AdvancedTuning {
     pub enable_post_grow_bfs_regrow: bool,
     /// When `true` AND [`enable_post_grow_refit`] triggered a refit,
     /// run a non-destructive cardinal-neighbour BFS extension
-    /// (`projective_grid::seed_and_grow::grow_extend::extend_from_labelled`) over
+    /// (`projective_grid::shared::grow_extend::extend_from_labelled`) over
     /// the existing labelled set with the refined centres. Walks the
     /// labelled bbox boundary one cell at a time, predicts each cell
     /// from cardinal labelled neighbours only (K=1 — much more
@@ -618,7 +618,7 @@ pub struct AdvancedTuning {
 
     // --- `extend_boundary` stage --------------------------------------------
     /// Use the per-candidate local-homography boundary extension
-    /// (`projective_grid::seed_and_grow::extension::extend_via_local_homography`)
+    /// (`projective_grid::shared::extension::extend_via_local_homography`)
     /// instead of the single-global-H one. The local-H variant fits an
     /// H per candidate cell from the K nearest labelled corners, gets
     /// per-candidate trust gates, and reaches further past the bbox
