+++ crates/projective-grid/src/detect.rs
@@ -1,12 +1,13 @@
 //! Detection task facade.
 //!
 //! The working implementations target the square lattice with any of the three
-//! input-feature kinds. [`Evidence::Oriented2`] is the native shape: a
-//! seed-and-grow pipeline (seed → BFS grow → validate → fit) or an axis-driven
-//! topological grid finder, selected via [`SquareAlgorithm`].
-//! [`Evidence::Positions`] (orientation-free) and [`Evidence::Oriented1`]
-//! (single-axis) are synthesized up to the Oriented2 shape
-//! ([`crate::orient`]) and then run the same strategies — so all three
+//! input-feature kinds. [`Evidence::Oriented2`] is the native shape, assembled
+//! by the axis-driven topological grid finder (Delaunay → quad-mesh →
+//! flood-fill → validate → fit). [`Evidence::Positions`] (orientation-free) and
+//! [`Evidence::Oriented1`] (single-axis) are synthesized up to the Oriented2
+//! shape ([`crate::orient`]) and then run the same assembler, with the
+//! geometry-only [`RecoverySchedule`] enabled to recover the recall the
+//! synthesized-axis frontier would otherwise leave on the table — so all three
 //! square input kinds share one back-half. All produce the same
 //! [`GridSolution`] shape.
 //!
@@ -23,30 +24,32 @@ use crate::feature::{CoordinateHypothesis, OrientedFeature, PointFeature};
 use crate::lattice::{GridDimensions, LatticeKind};
 use crate::result::{GridSolution, RejectedFeature};
 
-pub use crate::seed_and_grow::grow::GrowParams;
-pub use crate::seed_and_grow::recovery::{RecoveryParams, RecoverySchedule};
-pub use crate::seed_and_grow::seed::finder::SeedQuadParams as SeedParams;
+pub use crate::shared::recovery::{RecoveryParams, RecoverySchedule};
 pub use crate::shared::validate::ValidationParams as ValidateParams;
 pub use crate::topological::TopologicalParams;
 
 /// Algorithm selector for `(LatticeKind::Square, Evidence::Oriented2)`.
 ///
-/// Both variants accept the same [`Evidence::Oriented2`] input and
-/// produce the same [`GridSolution`] output; they differ in how they
-/// build the grid graph.
+/// The square grid is assembled by the [`Topological`](Self::Topological)
+/// axis-driven grid finder (the Shu/Brunton/Fiala 2009 image-free variant:
+/// Delaunay triangulation + axis-driven cell test + flood-fill + validate +
+/// fit). It is the only square assembler.
 ///
-/// * [`SeedAndGrow`](Self::SeedAndGrow) is the default mature square
-///   assembly path.
-/// * [`Topological`](Self::Topological) is the Shu/Brunton/Fiala 2009
-///   axis-driven grid finder. Image-free; tends to recover denser grids
-///   on clean inputs but is more sensitive to per-corner axis quality.
+/// The enum is retained as a single-variant, `#[non_exhaustive]` type so the
+/// public request shape (`detect_grid` / `detect_grid_all` /
+/// [`DetectionParams::algorithm`]) stays stable across the seed-and-grow
+/// retirement and a future alternative builder can be added without a breaking
+/// change. The historical `SeedAndGrow` variant — a self-consistent 4-corner
+/// seed plus BFS grow — was removed once the topological builder matched or
+/// beat it on every shipping path. Its geometry-only recovery schedule
+/// ([`RecoverySchedule`]) survives the retirement and now powers the
+/// topological synthesized-axis path.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
 #[non_exhaustive]
 pub enum SquareAlgorithm {
-    /// Seed-quad finder + BFS grow + shared validate + fit.
+    /// Axis-driven topological grid finder + shared validate + fit; the only
+    /// square assembler.
     #[default]
-    SeedAndGrow,
-    /// Axis-driven topological grid finder + shared validate + fit.
     Topological,
 }
 
@@ -100,35 +103,28 @@ impl Evidence<'_> {
 
 /// Detection parameters.
 ///
-/// The single `max_residual_px` knob from the Phase-A shell sits alongside
-/// per-algorithm sub-configs. The
-/// [`algorithm`](Self::algorithm) selector decides which of the
-/// `(Square, Oriented2)` paths consumes which sub-config; combinations
-/// not run by the chosen algorithm leave their sub-configs untouched.
+/// The single `max_residual_px` knob from the Phase-A shell sits alongside the
+/// topological assembler's sub-config, the shared post-detection validation
+/// tuning, and the post-recovery schedule. The
+/// [`algorithm`](Self::algorithm) selector is retained for request-shape
+/// stability; it currently has a single value
+/// ([`SquareAlgorithm::Topological`]).
 ///
-/// The sub-config types (`SeedParams`, `GrowParams`, `ValidateParams`)
-/// are the advanced-engine configs and do not implement `PartialEq`, so
-/// neither does `DetectionParams`.
+/// The sub-config types (`TopologicalParams`, `ValidateParams`,
+/// `RecoveryParams`) include advanced-engine configs that do not implement
+/// `PartialEq`, so neither does `DetectionParams`.
 #[derive(Clone, Debug)]
 #[non_exhaustive]
 pub struct DetectionParams {
     /// Residual threshold in image pixels for algorithms that fit a lattice.
     pub max_residual_px: f32,
     /// Algorithm picker for `(Square, Oriented2)`. Defaults to
-    /// [`SquareAlgorithm::SeedAndGrow`] so callers compile without
-    /// requiring an explicit algorithm choice.
+    /// [`SquareAlgorithm::Topological`] (the only square assembler); retained
+    /// for request-shape stability.
     pub algorithm: SquareAlgorithm,
-    /// Seed-quad finder tuning — consumed iff `algorithm ==
-    /// SquareAlgorithm::SeedAndGrow`.
-    pub seed: SeedParams,
-    /// BFS grow engine tuning — consumed iff `algorithm ==
-    /// SquareAlgorithm::SeedAndGrow`.
-    pub grow: GrowParams,
-    /// Topological grid-finder tuning — consumed iff `algorithm ==
-    /// SquareAlgorithm::Topological`.
+    /// Topological grid-finder tuning.
     pub topological: TopologicalParams,
-    /// Post-detection validation tuning — shared between both
-    /// `(Square, Oriented2)` algorithm paths.
+    /// Post-detection validation tuning.
     pub validate: ValidateParams,
     /// Post-convergence recovery schedule. Defaults to
     /// [`RecoverySchedule::Auto`]: the facade runs the geometry-only recovery
@@ -144,8 +140,6 @@ impl Default for DetectionParams {
         Self {
             max_residual_px: 2.0,
             algorithm: SquareAlgorithm::default(),
-            seed: SeedParams::default(),
-            grow: GrowParams::default(),
             topological: TopologicalParams::default(),
             validate: ValidateParams::default(),
             recovery: RecoverySchedule::default(),
@@ -163,8 +157,8 @@ impl DetectionParams {
         }
     }
 
-    /// Builder-style override: pick a different `(Square, Oriented2)`
-    /// algorithm. Default is [`SquareAlgorithm::SeedAndGrow`].
+    /// Builder-style override: pick the `(Square, Oriented2)` algorithm.
+    /// Default is [`SquareAlgorithm::Topological`] (the only square assembler).
     pub fn with_algorithm(mut self, algorithm: SquareAlgorithm) -> Self {
         self.algorithm = algorithm;
         self
@@ -179,18 +173,6 @@ impl DetectionParams {
     /// Builder-style override: replace the validate sub-config.
     pub fn with_validate(mut self, validate: ValidateParams) -> Self {
         self.validate = validate;
-        self
-    }
-
-    /// Builder-style override: replace the seed-quad sub-config.
-    pub fn with_seed(mut self, seed: SeedParams) -> Self {
-        self.seed = seed;
-        self
-    }
-
-    /// Builder-style override: replace the BFS-grow sub-config.
-    pub fn with_grow(mut self, grow: GrowParams) -> Self {
-        self.grow = grow;
         self
     }
 
@@ -244,35 +226,30 @@ impl<'a> DetectionRequest<'a> {
 ///
 /// | `(lattice, evidence)` | Status |
 /// |---|---|
-/// | `(Square, Oriented2)` | supported — two algorithms |
+/// | `(Square, Oriented2)` | supported — topological assembler |
 /// | `(Square, Oriented1)` | supported — synthesize 2nd axis, then Oriented2 |
 /// | `(Square, Positions)` | supported — synthesize both axes, then Oriented2 |
 /// | `(Square, Oriented3)` | `UnsupportedCombination` |
 /// | `(Square, CoordinateHypotheses)` | `UnsupportedCombination` (roadmap) |
-/// | `(Hex, Oriented3)` | supported — topological only (`algorithm = Topological`) |
+/// | `(Hex, Oriented3)` | supported — topological only |
 /// | `(Hex, Positions)` | supported — synthesize 3 axes, then hex topological |
 /// | `(Hex, Oriented1 / Oriented2)` | `UnsupportedCombination` |
-/// | `(Hex, *)` with `algorithm = SeedAndGrow` | `UnsupportedCombination` |
 ///
-/// * `(Square, Oriented2)` — two algorithm choices, picked via
-///   [`DetectionParams::algorithm`]:
-///     - [`SquareAlgorithm::SeedAndGrow`] (default): seed-quad finder +
-///       BFS grow + validate + fit.
-///     - [`SquareAlgorithm::Topological`]: axis-driven SBF09 grid finder +
-///       validate + fit.
-///
-///   Both return a labelled [`GridSolution`] with a fitted projective
-///   transform; downstream consumers stay agnostic.
+/// * `(Square, Oriented2)` — the [`SquareAlgorithm::Topological`] axis-driven
+///   SBF09 grid finder (Delaunay → quad-mesh → flood-fill → validate → fit)
+///   returns a labelled [`GridSolution`] with a fitted projective transform;
+///   downstream consumers stay agnostic.
 /// * `(Square, Positions)` — orientation-free input. Each corner's two
 ///   local grid directions are synthesized from neighbour geometry
-///   ([`crate::orient::synthesize_oriented2`]) and then fed to the
-///   algorithm picked by [`DetectionParams::algorithm`], exactly as for
-///   `(Square, Oriented2)`. Use this for dot / circle grids and for
-///   chessboards whose corners carry no axis estimate.
+///   ([`crate::orient::synthesize_oriented2`]) and then fed to the topological
+///   assembler, exactly as for `(Square, Oriented2)` — with the geometry-only
+///   [`RecoverySchedule`] enabled to recover the synthesized-axis recall.
+///   Use this for dot / circle grids and for chessboards whose corners carry
+///   no axis estimate.
 /// * `(Square, Oriented1)` — single-axis input. The supplied axis is kept
 ///   and the orthogonal grid direction is recovered from neighbour geometry
 ///   ([`crate::orient::synthesize_oriented2_from_oriented1`]); the resulting
-///   [`OrientedFeature<2>`] then runs the chosen algorithm, exactly as for
+///   [`OrientedFeature<2>`] then runs the topological assembler, exactly as for
 ///   `(Square, Positions)`. Use this for detectors that recover one dominant
 ///   edge orientation per feature but not the orthogonal one.
 /// * Every other combination — typed [`GridError::UnsupportedCombination`].
@@ -280,27 +257,22 @@ impl<'a> DetectionRequest<'a> {
 /// * `(Hex, Oriented3)` — hex-native triple-axis evidence. Runs the hex
 ///   topological grid finder (Delaunay triangles *are* the unit cells; no
 ///   diagonal class, no triangle-pair merge; axial `(q, r)` flood-fill walk).
-///   Hex is **topological-only** with **no recovery schedule**, so set
-///   [`DetectionParams::algorithm`] to [`SquareAlgorithm::Topological`]; the
-///   default [`SquareAlgorithm::SeedAndGrow`] yields `UnsupportedCombination`
-///   for any hex request (no seed-and-grow hex implementation).
+///   Hex is **topological-only** with **no recovery schedule**.
 /// * `(Hex, Positions)` — orientation-free hex input. The three local grid
 ///   directions are synthesized from neighbour geometry
 ///   ([`crate::orient::synthesize_oriented3`]) and then fed to the hex
 ///   topological path, mirroring the `(Square, Positions)` seam.
 ///
 /// `(Square, Oriented3)` (square does not consume triple-axis evidence),
-/// `(Square, CoordinateHypotheses)` (a decode-feedback roadmap slot),
-/// `(Hex, Oriented1)` / `(Hex, Oriented2)` (hex needs three axis families),
-/// and any `(Hex, *)` request with `algorithm = SeedAndGrow` stay
-/// `UnsupportedCombination` — no working algorithm exists for those slots.
+/// `(Square, CoordinateHypotheses)` (a decode-feedback roadmap slot), and
+/// `(Hex, Oriented1)` / `(Hex, Oriented2)` (hex needs three axis families)
+/// stay `UnsupportedCombination` — no working algorithm exists for those slots.
 ///
-/// **Multi-component results.** Both algorithms can produce more than one
-/// connected component (seed-and-grow assembles each disconnected patch
-/// independently, then runs local component merge; the topological path
-/// labels each connected quad-mesh component). This entry point returns
-/// the largest component only. Use [`detect_grid_all`] when secondary
-/// components must be preserved with their own `(u, v)` labels.
+/// **Multi-component results.** The topological assembler can produce more
+/// than one connected component (it labels each connected quad-mesh component,
+/// then runs local component merge). This entry point returns the largest
+/// component only. Use [`detect_grid_all`] when secondary components must be
+/// preserved with their own `(u, v)` labels.
 pub fn detect_grid(request: DetectionRequest<'_>) -> Result<GridSolution> {
     let mut report = detect_grid_all(request)?;
     if report.solutions.is_empty() {
@@ -311,21 +283,15 @@ pub fn detect_grid(request: DetectionRequest<'_>) -> Result<GridSolution> {
 }
 
 /// Dispatch oriented-2 features (caller-supplied or synthesized) to the
-/// algorithm picked by [`DetectionParams::algorithm`]. The single dispatch
-/// point shared by the `Oriented2`, `Positions`, and `Oriented1` arms so the
-/// three input kinds reach identical strategy code.
+/// topological square assembler. The single dispatch point shared by the
+/// `Oriented2`, `Positions`, and `Oriented1` arms so the three input kinds
+/// reach identical strategy code.
 fn run_square_oriented2(
     features: &[OrientedFeature<2>],
     request: &DetectionRequest<'_>,
     synthesized_axes: bool,
 ) -> Result<Vec<GridSolution>> {
     match request.params.algorithm {
-        SquareAlgorithm::SeedAndGrow => crate::seed_and_grow::detect_square_oriented2_seed_grow(
-            features,
-            request.dimensions,
-            &request.params,
-            synthesized_axes,
-        ),
         SquareAlgorithm::Topological => {
             crate::topological::detect_square_oriented2_topological_all(
                 features,
@@ -340,14 +306,10 @@ fn run_square_oriented2(
 /// Dispatch hex triple-axis features (caller-supplied or synthesized) to the
 /// hex topological path.
 ///
-/// Hex detection is **topological-only**: the
-/// [`SquareAlgorithm::SeedAndGrow`] selector has no hex implementation (the
-/// seed-and-grow recovery machinery is square/ChESS-axis coupled), so a hex
-/// request with `algorithm == SeedAndGrow` is a typed
-/// [`GridError::UnsupportedCombination`] rather than a silent fallback. The
-/// algorithm selector enum is named [`SquareAlgorithm`]; for hex only its
-/// `Topological` variant is meaningful, and it is the default-equivalent path
-/// here (see the support matrix on [`detect_grid`]).
+/// Hex detection is **topological-only**. The algorithm selector enum is named
+/// [`SquareAlgorithm`]; its sole [`Topological`](SquareAlgorithm::Topological)
+/// variant is the default-equivalent path here (see the support matrix on
+/// [`detect_grid`]).
 fn run_hex_oriented3(
     features: &[OrientedFeature<3>],
     request: &DetectionRequest<'_>,
@@ -358,11 +320,6 @@ fn run_hex_oriented3(
             request.dimensions,
             &request.params,
         ),
-        SquareAlgorithm::SeedAndGrow => Err(GridError::UnsupportedCombination {
-            task: GridTask::Detection,
-            lattice: LatticeKind::Hex,
-            evidence: request.evidence.kind(),
-        }),
     }
 }
 
@@ -370,9 +327,8 @@ fn run_hex_oriented3(
 ///
 /// Returns a [`DetectionReport`] with one [`GridSolution`] per
 /// qualifying connected component, ordered by labelled-count
-/// descending (ties broken by smallest labelled `source_index`). Both
-/// the seed-and-grow and topological algorithms may return several
-/// solutions.
+/// descending (ties broken by smallest labelled `source_index`). The
+/// topological assembler may return several solutions.
 ///
 /// Features that no component admitted are surfaced in the *first*
 /// solution's `rejected` vector (the same shape callers of
@@ -408,8 +364,7 @@ pub fn detect_grid_all(request: DetectionRequest<'_>) -> Result<DetectionReport>
         }
         (LatticeKind::Hex, Evidence::Oriented3(features)) => {
             // Hex-native triple-axis evidence. Hex detection is
-            // topological-only; seed-and-grow hex is a typed
-            // `UnsupportedCombination` below.
+            // topological-only.
             run_hex_oriented3(features, &request)?
         }
         (LatticeKind::Hex, Evidence::Positions(features)) => {
@@ -438,8 +393,8 @@ pub fn detect_grid_all(request: DetectionRequest<'_>) -> Result<DetectionReport>
 /// crate-level slot reserved for features that no component admitted
 /// when the orchestrator (rather than a particular component) is the
 /// authoritative source of that information. It is left empty for the
-/// two algorithms currently implemented — the per-component rejected
-/// vectors already cover the wire shape.
+/// topological assembler — the per-component rejected vectors already
+/// cover the wire shape.
 #[derive(Clone, Debug)]
 #[non_exhaustive]
 pub struct DetectionReport {
@@ -447,8 +402,8 @@ pub struct DetectionReport {
     /// descending.
     pub solutions: Vec<GridSolution>,
     /// Features that no component admitted, scoped to the orchestrator
-    /// (not a particular component). Currently empty for both
-    /// `SquareAlgorithm` variants — see the struct-level docs.
+    /// (not a particular component). Currently empty for the topological
+    /// assembler — see the struct-level docs.
     pub rejected: Vec<RejectedFeature>,
 }
