+++ crates/projective-grid/src/shared/extension/local.rs
@@ -11,9 +11,9 @@ use kiddo::{KdTree, SquaredEuclidean};
 use nalgebra::Point2;
 
 use crate::geometry::estimate_homography;
-use crate::seed_and_grow::extension::common::{try_attach_at_cell, TryCellResult};
-use crate::seed_and_grow::extension::{ExtensionStats, LocalExtensionParams};
-use crate::seed_and_grow::grow::{GrowResult, SquareAttachPolicy};
+use crate::shared::extension::common::{try_attach_at_cell, TryCellResult};
+use crate::shared::extension::{ExtensionStats, LocalExtensionParams};
+use crate::shared::grow::{GrowResult, SquareAttachPolicy};
 
 /// Extend the labelled grid outward (and into interior holes) using a
 /// **per-candidate local homography** fit from the K nearest labelled
