+++ crates/projective-grid/tests/detect_placeholders.rs
@@ -28,43 +28,20 @@ fn assert_unsupported(request: DetectionRequest<'_>, evidence: EvidenceKind) {
 // success path is covered by `tests/detect_square_positions.rs`.
 //
 // `square_oriented_detection_is_typed_unsupported` was removed in Phase C:
-// `(LatticeKind::Square, Evidence::Oriented2)` now runs the seed-and-grow port
+// `(LatticeKind::Square, Evidence::Oriented2)` now runs the topological port
 // and returns a real labelled grid. The success path is covered by
 // `tests/detect_square_oriented2.rs`.
-
-#[test]
-fn hex_position_detection_is_typed_unsupported() {
-    let features = [point(0), point(1), point(2), point(3)];
-    let request = DetectionRequest::new(
-        LatticeKind::Hex,
-        Evidence::Positions(&features),
-        None,
-        DetectionParams::default(),
-    );
-    assert_unsupported(request, EvidenceKind::Positions);
-}
-
-#[test]
-fn hex_oriented_detection_is_typed_unsupported() {
-    let axis = LocalAxis::new(0.0_f32, None);
-    let features = [
-        OrientedFeature::<3>::new(
-            point(0),
-            [axis, LocalAxis::new(1.0, None), LocalAxis::new(2.0, None)],
-        ),
-        OrientedFeature::<3>::new(
-            point(1),
-            [axis, LocalAxis::new(1.0, None), LocalAxis::new(2.0, None)],
-        ),
-    ];
-    let request = DetectionRequest::new(
-        LatticeKind::Hex,
-        Evidence::Oriented3(&features),
-        None,
-        DetectionParams::default(),
-    );
-    assert_unsupported(request, EvidenceKind::Oriented3);
-}
+//
+// `hex_position_detection_is_typed_unsupported` and
+// `hex_oriented_detection_is_typed_unsupported` were removed with the
+// seed-and-grow retirement: they only asserted "unsupported" because the
+// historical default algorithm was `SeedAndGrow`, which had no hex path. With
+// the topological default, `(Hex, Positions)` and `(Hex, Oriented3)` are the
+// supported hex paths and route to the real hex topological detector (they
+// return `InsufficientEvidence` / `DegenerateGeometry` on degenerate input,
+// not `UnsupportedCombination`). The hex success path is covered by
+// `tests/detect_hex_positions.rs`; the remaining genuine hex-unsupported cases
+// — `(Hex, Oriented1)` / `(Hex, Oriented2)` — are covered there too.
 
 #[test]
 fn square_oriented3_detection_is_typed_unsupported() {
