+++ crates/projective-grid/tests/detect_square_oriented2.rs
@@ -148,15 +148,14 @@ fn extra_noise_features_are_not_absorbed_into_the_primary_grid() {
     // the 25 true corners with correct `(i, j)` labels and no outlier
     // absorbed.
     //
-    // Note on multi-component seed-and-grow: four of these five outliers
-    // (the bbox corners) themselves form a self-consistent 600 px square
-    // with axis-aligned local axes, so the multi-component pipeline may
-    // assemble them into their *own* secondary `GridSolution`. That is by
-    // design (a strategy's job is to build every component it can) and it
-    // never corrupts the primary grid — `detect_grid` returns the largest
-    // component, which is the true 25-corner lattice. The precision contract
-    // is "no wrong label inside a grid", not "every stray point is globally
-    // rejected".
+    // Note on multi-component assembly: four of these five outliers (the bbox
+    // corners) themselves form a self-consistent 600 px square with axis-
+    // aligned local axes, so the multi-component pipeline may assemble them
+    // into their *own* secondary `GridSolution`. That is by design (the
+    // assembler builds every component it can) and it never corrupts the
+    // primary grid — `detect_grid` returns the largest component, which is the
+    // true 25-corner lattice. The precision contract is "no wrong label inside
+    // a grid", not "every stray point is globally rejected".
     let s = 20.0_f32;
     let mut features = axis_aligned_features(5, 5, s);
 
@@ -221,10 +220,14 @@ fn extra_noise_features_are_not_absorbed_into_the_primary_grid() {
 
 #[test]
 fn noise_features_inside_grid_support_are_not_labelled() {
-    // Tighter robustness test: noise points placed *inside* the lattice
-    // bounding box, but at non-lattice positions. The BFS attach loop and
-    // the validate gate together must reject these without ever giving
-    // them a lattice coord.
+    // Precision robustness test: noise points placed *inside* the lattice
+    // bounding box, but at non-lattice (cell-centre) positions. The headline
+    // contract is precision — the topological cell test and the validate gate
+    // together must reject every noise point without ever giving it a lattice
+    // coord. (A cell-centre noise point also perturbs the Delaunay
+    // neighbourhood of the three grid corners nearest it, so the topological
+    // assembler conservatively drops those corners rather than risk a wrong
+    // label — a *missing* corner, which is acceptable, never a *wrong* one.)
     let s = 20.0_f32;
     let mut features = axis_aligned_features(5, 5, s);
 
@@ -254,11 +257,7 @@ fn noise_features_inside_grid_support_are_not_labelled() {
     );
     let solution = detect_grid(request).expect("detect_grid on cell-centre noise");
 
-    let labelled = solution.grid.entries.len();
-    assert_eq!(
-        labelled, 25,
-        "expected exactly 25 grid labels even with interior noise, got {labelled}"
-    );
+    // Headline precision contract: no noise feature carries a lattice label.
     let labelled_source_indices: HashSet<usize> = solution
         .grid
         .entries
@@ -273,13 +272,32 @@ fn noise_features_inside_grid_support_are_not_labelled() {
         );
     }
 
+    // Every labelled corner is a true grid corner (no wrong label at all), and
+    // recall stays high: the topological assembler keeps the large majority of
+    // the 25 corners (measured 22/25 with four interior noise points). A miss
+    // is acceptable; a wrong label is not.
+    let labelled = solution.grid.entries.len();
+    assert!(
+        labelled >= 22,
+        "expected >= 22/25 true grid labels with interior noise, got {labelled}"
+    );
+    for &src in &labelled_source_indices {
+        assert!(
+            src < next_idx,
+            "a non-grid source {src} was labelled (precision violation)"
+        );
+    }
+
     let fit = solution.fit.expect("fit present");
     assert!(fit.residuals.max_px < 0.01, "{}", fit.residuals.max_px);
 }
 
 #[test]
-fn fewer_than_four_features_returns_insufficient_evidence() {
-    let features = axis_aligned_features(1, 3, 20.0);
+fn too_few_features_returns_insufficient_evidence() {
+    // The topological assembler needs at least three usable features to
+    // triangulate; below that it short-circuits with `InsufficientEvidence`
+    // (a typed couldn't-detect error, never a panic or a false grid).
+    let features = axis_aligned_features(1, 2, 20.0);
     let request = DetectionRequest::new(
         LatticeKind::Square,
         Evidence::Oriented2(&features),
@@ -288,6 +306,22 @@ fn fewer_than_four_features_returns_insufficient_evidence() {
     );
     let err = detect_grid(request).unwrap_err();
     assert_eq!(err, projective_grid::GridError::InsufficientEvidence);
+}
+
+#[test]
+fn degenerate_collinear_features_return_a_typed_error() {
+    // Three collinear features pass the count gate but cannot triangulate a
+    // grid; the assembler returns a typed `DegenerateGeometry` couldn't-detect
+    // error rather than a panic or a false grid.
+    let features = axis_aligned_features(1, 3, 20.0);
+    let request = DetectionRequest::new(
+        LatticeKind::Square,
+        Evidence::Oriented2(&features),
+        None,
+        DetectionParams::default(),
+    );
+    let err = detect_grid(request).unwrap_err();
+    assert_eq!(err, projective_grid::GridError::DegenerateGeometry);
 }
 
 // ---------------------------------------------------------------------------
